TRINITY: another try to get this 64bit stuff to compile
Binaries submitted in 972227
Build time : 9 minutes 25.213000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 972208]

+++ Eve/SpaceObject/Attachments/EveImpactOverlay.cpp
@@ -71,7 +71,7 @@ void EveImpactOverlay::UpdateSyncronous( EveUpdateContext& updateContext, EveSpa
 	header.v[3] = Vector4( 0.f, 0.f, 0.f, 0.f );
 
 	// update block data
-	m_dataTextureBlockID = dataTextureMgr->RequestBlockData( &header.v[0], m_impactTexelData.size(), m_impactTexelData.empty() ? nullptr : &m_impactTexelData[0].v[0] );
+	m_dataTextureBlockID = dataTextureMgr->RequestBlockData( &header.v[0], (uint32_t)m_impactTexelData.size(), m_impactTexelData.empty() ? nullptr : &m_impactTexelData[0].v[0] );
 }
 
 // --------------------------------------------------------------------------------
